requirements.txt
faiss-cpu
pypdf
sentence-transformers
langchain
langchain_text_splitters
google-generativeai
streamlit